// src/types.ts
export interface Project {
    github: string;
    live: string;
    title: string;
    image: string;
}

export interface User {
    username: string;
    email: string;
    name: string;
    profilePic: string;
    profession: string;
    about: string;
    githubUrl: string;
    linkedinUrl: string;
    websiteUrl: string;
}